+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.0.0
+ 
+- Add support for Web platform
+- Add new class `ConnectionEntry` which is used to determine the connection params (address or api url)
+- Breaking changes: Now the package is using `ConnectionEntry` instead of `InternetAddress` to determine the connection params. This is because `InternetAddress` is not supported on Web.
+
 ## 1.1.1
 
 - Fix the problem where `connectionStream` wasn't emitting the initial event when being offline from the startup. Problem was faced on the Android platform.